refactor: move p5 highlight logic into p5JavaScript.js
